huffman: optimize dictionary size

+++ huffman.hpp
@@ -595,9 +595,9 @@ private:
 public:
 	/**
 		 * File Structure
-		 * Header: Dictionary Index Count(uint32_t)
-		 * Dictionary Header: Character(int8_t), SizeofData(uint32_t)
-		 * Dictionary Data:   Size(uint32_t), Align(uint8_t), Data
+		 * Header: Dictionary Index Count(uint16_t)
+		 * Dictionary Header: Character(int8_t), SizeofData(uint16_t)
+		 * Dictionary Data:   Size(uint16_t), Align(uint8_t), Data
 		 * File Data:         Size(uint32_t), Align(uint8_t), Data
 		*/
 	static void compress(const std::string &in, const std::string &out)
@@ -625,10 +625,10 @@ public:
 		}
 		int align = align_data(buff);
 		std::ofstream ofs(out, std::ios_base::binary);
-		write_data<std::uint32_t>(ofs, encode.size());
+		write_data<std::uint16_t>(ofs, encode.size());
 		for (auto &it : encode) {
 			write_data<std::int8_t>(ofs, it.first);
-			write_data<std::uint32_t>(ofs, it.second.size());
+			write_data<std::uint16_t>(ofs, it.second.size());
 		}
 		std::vector<char> dict_buff;
 		for (auto &it : encode) {
@@ -636,7 +636,7 @@ public:
 				dict_buff.push_back(ch);
 		}
 		int dict_align = align_data(dict_buff);
-		write_data<std::uint32_t>(ofs, dict_buff.size() / 8);
+		write_data<std::uint16_t>(ofs, dict_buff.size() / 8);
 		write_data<std::uint8_t>(ofs, dict_align);
 		write_data(ofs, dict_buff);
 		write_data<std::uint32_t>(ofs, buff.size() / 8);
@@ -648,16 +648,16 @@ public:
 		std::ifstream ifs(in, std::ios_base::binary);
         if (!ifs)
             throw std::runtime_error("File not exist.");
-		std::size_t count = read_data<std::uint32_t>(ifs);
+		std::size_t count = read_data<std::uint16_t>(ifs);
 		std::vector<std::pair<char, std::size_t>> index;
 		for (std::size_t i = 0; i < count; ++i) {
 			char ch = read_data<std::int8_t>(ifs);
-			std::size_t size = read_data<std::uint32_t>(ifs);
+			std::size_t size = read_data<std::uint16_t>(ifs);
 			index.emplace_back(ch, size);
 		}
 		std::vector<char> buff;
 		{
-			int data_size = read_data<std::uint32_t>(ifs);
+			int data_size = read_data<std::uint16_t>(ifs);
 			int align = read_data<std::uint8_t>(ifs);
 			for (std::size_t i = 0; i < data_size; ++i)
 				buff.push_back(ifs.get());
